@@ -5,7 +5,7 @@ version: 3.0.2
 homepage: https://github.com/De-Morgan/flutter-contextual-action-bar
 
 environment:
-  sdk: '>=2.17.6 <3.0.0'
+  sdk: ^3.0.0
 
 dependencies:
   flutter:
