@@ -1,31 +1,6 @@
 part of widgets;
 
-class ContextualAppBar<T> extends StatefulWidget
-    implements PreferredSizeWidget {
-  ContextualAppBar({
-    Key? key,
-    required this.counterBuilder,
-    required this.contextualActions,
-    this.flexibleSpace,
-    this.bottom,
-    this.elevation,
-    this.shape,
-    this.backgroundColor,
-    this.brightness,
-    this.iconTheme,
-    this.actionsIconTheme,
-    this.textTheme,
-    this.closeIcon,
-    this.primary = true,
-    this.centerTitle,
-    this.titleSpacing = NavigationToolbar.kMiddleSpacing,
-    this.toolbarOpacity = 1.0,
-    this.bottomOpacity = 1.0,
-  })  : assert(elevation == null || elevation >= 0.0),
-        preferredSize = Size.fromHeight(
-            kToolbarHeight + (bottom?.preferredSize.height ?? 0.0)),
-        super(key: key);
-
+class ContextualAppBar<T> extends StatelessWidget {
   /// The primary widget displayed in the app bar.
   ///  The itemsCount property gives you the count of the selected items.
   /// Typically a [Text] widget containing a description of the current contents
@@ -39,99 +14,261 @@ class ContextualAppBar<T> extends StatefulWidget
   /// last action.
   final List<ContextualAction<T>> contextualActions;
 
-  /// This widget is stacked behind the toolbar and the tab bar. It's height will
+  /// {@template flutter.material.appbar.automaticallyImplyLeading}
+  /// Controls whether we should try to imply the leading widget if null.
+  ///
+  /// If true and [leading] is null, automatically try to deduce what the leading
+  /// widget should be. If false and [leading] is null, leading space is given to [title].
+  /// If leading widget is not null, this parameter has no effect.
+  /// {@endtemplate}
+  final bool automaticallyImplyLeading;
+
+  /// Icon to use to close Action mode
+  final IconData? closeIcon;
+
+  /// {@template flutter.material.appbar.flexibleSpace}
+  /// This widget is stacked behind the toolbar and the tab bar. Its height will
   /// be the same as the app bar's overall height.
   ///
-  /// A flexible space isn't actually flexible unless the [ContextualAppBar]'s container
-  /// changes the [ContextualAppBar]'s size. A [SliverAppBar] in a [CustomScrollView]
-  /// changes the [ContextualAppBar]'s height when scrolled.
+  /// A flexible space isn't actually flexible unless the [AppBar]'s container
+  /// changes the [AppBar]'s size. A [SliverAppBar] in a [CustomScrollView]
+  /// changes the [AppBar]'s height when scrolled.
   ///
   /// Typically a [FlexibleSpaceBar]. See [FlexibleSpaceBar] for details.
+  /// {@endtemplate}
   final Widget? flexibleSpace;
 
+  /// {@template flutter.material.appbar.bottom}
   /// This widget appears across the bottom of the app bar.
   ///
   /// Typically a [TabBar]. Only widgets that implement [PreferredSizeWidget] can
   /// be used at the bottom of an app bar.
+  /// {@endtemplate}
   ///
   /// See also:
   ///
   ///  * [PreferredSize], which can be used to give an arbitrary widget a preferred size.
   final PreferredSizeWidget? bottom;
 
+  /// {@template flutter.material.appbar.elevation}
   /// The z-coordinate at which to place this app bar relative to its parent.
   ///
-  /// This controls the size of the shadow below the app bar.
+  /// This property controls the size of the shadow below the app bar if
+  /// [shadowColor] is not null.
   ///
-  /// The value is non-negative.
+  /// If [surfaceTintColor] is not null then it will apply a surface tint overlay
+  /// to the background color (see [Material.surfaceTintColor] for more
+  /// detail).
   ///
-  /// If this property is null, then [ThemeData.appBarTheme.elevation] is used,
-  /// if that is also null, the default value is 4, the appropriate elevation
-  /// for app bars.
+  /// The value must be non-negative.
+  ///
+  /// If this property is null, then [AppBarTheme.elevation] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, the
+  /// default value is 4.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [scrolledUnderElevation], which will be used when the app bar has
+  ///    something scrolled underneath it.
+  ///  * [shadowColor], which is the color of the shadow below the app bar.
+  ///  * [surfaceTintColor], which determines the elevation overlay that will
+  ///    be applied to the background of the app bar.
+  ///  * [shape], which defines the shape of the app bar's [Material] and its
+  ///    shadow.
   final double? elevation;
 
-  /// The material's shape as well its shadow.
+  /// {@template flutter.material.appbar.scrolledUnderElevation}
+  /// The elevation that will be used if this app bar has something
+  /// scrolled underneath it.
+  ///
+  /// If non-null then it [AppBarTheme.scrolledUnderElevation] of
+  /// [ThemeData.appBarTheme] will be used. If that is also null then [elevation]
+  /// will be used.
+  ///
+  /// The value must be non-negative.
+  ///
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///  * [elevation], which will be used if there is no content scrolled under
+  ///    the app bar.
+  ///  * [shadowColor], which is the color of the shadow below the app bar.
+  ///  * [surfaceTintColor], which determines the elevation overlay that will
+  ///    be applied to the background of the app bar.
+  ///  * [shape], which defines the shape of the app bar's [Material] and its
+  ///    shadow.
+  final double? scrolledUnderElevation;
+
+  /// A check that specifies which child's [ScrollNotification]s should be
+  /// listened to.
+  ///
+  /// By default, checks whether `notification.depth == 0`. Set it to something
+  /// else for more complicated layouts.
+  final ScrollNotificationPredicate notificationPredicate;
+
+  /// {@template flutter.material.appbar.shadowColor}
+  /// The color of the shadow below the app bar.
+  ///
+  /// If this property is null, then [AppBarTheme.shadowColor] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, the default value
+  /// is fully opaque black.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [elevation], which defines the size of the shadow below the app bar.
+  ///  * [shape], which defines the shape of the app bar and its shadow.
+  final Color? shadowColor;
+
+  /// {@template flutter.material.appbar.surfaceTintColor}
+  /// The color of the surface tint overlay applied to the app bar's
+  /// background color to indicate elevation.
+  ///
+  /// If null no overlay will be applied.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///   * [Material.surfaceTintColor], which described this feature in more detail.
+  final Color? surfaceTintColor;
+
+  /// {@template flutter.material.appbar.shape}
+  /// The shape of the app bar's [Material] as well as its shadow.
+  ///
+  /// If this property is null, then [AppBarTheme.shape] of
+  /// [ThemeData.appBarTheme] is used. Both properties default to null.
+  /// If both properties are null then the shape of the app bar's [Material]
+  /// is just a simple rectangle.
   ///
   /// A shadow is only displayed if the [elevation] is greater than
   /// zero.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [elevation], which defines the size of the shadow below the app bar.
+  ///  * [shadowColor], which is the color of the shadow below the app bar.
   final ShapeBorder? shape;
 
-  /// The color to use for the app bar's material. Typically this should be set
-  /// along with [brightness], [iconTheme], [textTheme].
+  /// {@template flutter.material.appbar.backgroundColor}
+  /// The fill color to use for an app bar's [Material].
   ///
-  /// If this property is null, then [ThemeData.appBarTheme.color] is used,
-  /// if that is also null, then [ThemeData.primaryColor] is used.
+  /// If null, then the [AppBarTheme.backgroundColor] is used. If that value is also
+  /// null, then [AppBar] uses the overall theme's [ColorScheme.primary] if the
+  /// overall theme's brightness is [Brightness.light], and [ColorScheme.surface]
+  /// if the overall theme's brightness is [Brightness.dark].
+  ///
+  /// If this color is a [MaterialStateColor] it will be resolved against
+  /// [MaterialState.scrolledUnder] when the content of the app's
+  /// primary scrollable overlaps the app bar.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [foregroundColor], which specifies the color for icons and text within
+  ///    the app bar.
+  ///  * [Theme.of], which returns the current overall Material theme as
+  ///    a [ThemeData].
+  ///  * [ThemeData.colorScheme], the thirteen colors that most Material widget
+  ///    default colors are based on.
+  ///  * [ColorScheme.brightness], which indicates if the overall [Theme]
+  ///    is light or dark.
   final Color? backgroundColor;
 
-  /// The brightness of the app bar's material. Typically this is set along
-  /// with [backgroundColor], [iconTheme], [textTheme].
+  /// {@template flutter.material.appbar.foregroundColor}
+  /// The default color for [Text] and [Icon]s within the app bar.
   ///
-  /// If this property is null, then [ThemeData.appBarTheme.brightness] is used,
-  /// if that is also null, then [ThemeData.primaryColorBrightness] is used.
-  final Brightness? brightness;
+  /// If null, then [AppBarTheme.foregroundColor] is used. If that
+  /// value is also null, then [AppBar] uses the overall theme's
+  /// [ColorScheme.onPrimary] if the overall theme's brightness is
+  /// [Brightness.light], and [ColorScheme.onSurface] if the overall
+  /// theme's brightness is [Brightness.dark].
+  ///
+  /// This color is used to configure [DefaultTextStyle] that contains
+  /// the toolbar's children, and the default [IconTheme] widgets that
+  /// are created if [iconTheme] and [actionsIconTheme] are null.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [backgroundColor], which specifies the app bar's background color.
+  ///  * [Theme.of], which returns the current overall Material theme as
+  ///    a [ThemeData].
+  ///  * [ThemeData.colorScheme], the thirteen colors that most Material widget
+  ///    default colors are based on.
+  ///  * [ColorScheme.brightness], which indicates if the overall [Theme]
+  ///    is light or dark.
+  final Color? foregroundColor;
 
-  /// The color, opacity, and size to use for app bar icons. Typically this
-  /// is set along with [backgroundColor], [brightness], [textTheme].
+  /// {@template flutter.material.appbar.iconTheme}
+  /// The color, opacity, and size to use for toolbar icons.
   ///
-  /// If this property is null, then [ThemeData.appBarTheme.iconTheme] is used,
-  /// if that is also null, then [ThemeData.primaryIconTheme] is used.
+  /// If this property is null, then a copy of [ThemeData.iconTheme]
+  /// is used, with the [IconThemeData.color] set to the
+  /// app bar's [foregroundColor].
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [actionsIconTheme], which defines the appearance of icons in
+  ///    the [actions] list.
   final IconThemeData? iconTheme;
 
+  /// {@template flutter.material.appbar.actionsIconTheme}
   /// The color, opacity, and size to use for the icons that appear in the app
-  /// bar's [contextualActions]. This should only be used when the [contextualActions] should be
+  /// bar's [actions].
+  ///
+  /// This property should only be used when the [actions] should be
   /// themed differently than the icon that appears in the app bar's [leading]
   /// widget.
   ///
-  /// If this property is null, then [ThemeData.appBarTheme.actionsIconTheme] is
-  /// used, if that is also null, then this falls back to [iconTheme].
+  /// If this property is null, then [AppBarTheme.actionsIconTheme] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, then the value of
+  /// [iconTheme] is used.
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [iconTheme], which defines the appearance of all of the toolbar icons.
   final IconThemeData? actionsIconTheme;
 
-  /// The typographic styles to use for text in the app bar. Typically this is
-  /// set along with [brightness] [backgroundColor], [iconTheme].
-  ///
-  /// If this property is null, then [ThemeData.appBarTheme.textTheme] is used,
-  /// if that is also null, then [ThemeData.primaryTextTheme] is used.
-  final TextTheme? textTheme;
-
+  /// {@template flutter.material.appbar.primary}
   /// Whether this app bar is being displayed at the top of the screen.
   ///
   /// If true, the app bar's toolbar elements and [bottom] widget will be
   /// padded on top by the height of the system status bar. The layout
   /// of the [flexibleSpace] is not affected by the [primary] property.
+  /// {@endtemplate}
   final bool primary;
 
+  /// {@template flutter.material.appbar.centerTitle}
   /// Whether the title should be centered.
   ///
-  /// Defaults to being adapted to the current [TargetPlatform].
+  /// If this property is null, then [AppBarTheme.centerTitle] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, then value is
+  /// adapted to the current [TargetPlatform].
+  /// {@endtemplate}
   final bool? centerTitle;
 
-  /// The spacing around [counterBuilder] content on the horizontal axis. This spacing is
-  /// applied even if there is no [leading] content or [contextualActions]. If you want
-  /// [counterBuilder] to take all the space available, set this value to 0.0.
+  /// {@template flutter.material.appbar.excludeHeaderSemantics}
+  /// Whether the title should be wrapped with header [Semantics].
   ///
-  /// Defaults to [NavigationToolbar.kMiddleSpacing].
-  final double titleSpacing;
+  /// Defaults to false.
+  /// {@endtemplate}
+  final bool excludeHeaderSemantics;
 
+  /// {@template flutter.material.appbar.titleSpacing}
+  /// The spacing around [title] content on the horizontal axis. This spacing is
+  /// applied even if there is no [leading] content or [actions]. If you want
+  /// [title] to take all the space available, set this value to 0.0.
+  ///
+  /// If this property is null, then [AppBarTheme.titleSpacing] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, then the
+  /// default value is [NavigationToolbar.kMiddleSpacing].
+  /// {@endtemplate}
+  final double? titleSpacing;
+
+  /// {@template flutter.material.appbar.toolbarOpacity}
   /// How opaque the toolbar part of the app bar is.
   ///
   /// A value of 1.0 is fully opaque, and a value of 0.0 is fully transparent.
@@ -139,8 +276,10 @@ class ContextualAppBar<T> extends StatefulWidget
   /// Typically, this value is not changed from its default value (1.0). It is
   /// used by [SliverAppBar] to animate the opacity of the toolbar when the app
   /// bar is scrolled.
+  /// {@endtemplate}
   final double toolbarOpacity;
 
+  /// {@template flutter.material.appbar.bottomOpacity}
   /// How opaque the bottom part of the app bar is.
   ///
   /// A value of 1.0 is fully opaque, and a value of 0.0 is fully transparent.
@@ -148,272 +287,172 @@ class ContextualAppBar<T> extends StatefulWidget
   /// Typically, this value is not changed from its default value (1.0). It is
   /// used by [SliverAppBar] to animate the opacity of the toolbar when the app
   /// bar is scrolled.
+  /// {@endtemplate}
   final double bottomOpacity;
 
-  /// Icon to use to close Action mode
-  final IconData? closeIcon;
-
-  /// A size whose height is the sum of [kToolbarHeight] and the [bottom] widget's
-  /// preferred height.
+  /// {@template flutter.material.appbar.toolbarHeight}
+  /// Defines the height of the toolbar component of an [AppBar].
   ///
-  /// [Scaffold] uses this size to set its app bar's height.
-  @override
-  final Size preferredSize;
+  /// By default, the value of [toolbarHeight] is [kToolbarHeight].
+  /// {@endtemplate}
+  final double? toolbarHeight;
 
-  bool? _getEffectiveCenterTitle(ThemeData theme) {
-    if (centerTitle != null) return centerTitle;
-    switch (theme.platform) {
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-        return false;
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        return contextualActions.length < 2;
-      case TargetPlatform.linux:
-        break;
-      case TargetPlatform.windows:
-        break;
-    }
-    return null;
-  }
+  /// {@template flutter.material.appbar.leadingWidth}
+  /// Defines the width of [leading] widget.
+  ///
+  /// By default, the value of [leadingWidth] is 56.0.
+  /// {@endtemplate}
+  final double? leadingWidth;
 
-  @override
-  _ContextualAppBarState createState() => _ContextualAppBarState<T>();
-}
+  /// {@template flutter.material.appbar.toolbarTextStyle}
+  /// The default text style for the AppBar's [leading], and
+  /// [actions] widgets, but not its [title].
+  ///
+  /// If this property is null, then [AppBarTheme.toolbarTextStyle] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, the default
+  /// value is a copy of the overall theme's [TextTheme.bodyMedium]
+  /// [TextStyle], with color set to the app bar's [foregroundColor].
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [titleTextStyle], which overrides the default text style for the [title].
+  ///  * [DefaultTextStyle], which overrides the default text style for all of the
+  ///    widgets in a subtree.
+  final TextStyle? toolbarTextStyle;
 
-class _ContextualAppBarState<T> extends State<ContextualAppBar> {
-  static const double _defaultElevation = 4.0;
+  /// {@template flutter.material.appbar.titleTextStyle}
+  /// The default text style for the AppBar's [title] widget.
+  ///
+  /// If this property is null, then [AppBarTheme.titleTextStyle] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, the default
+  /// value is a copy of the overall theme's [TextTheme.titleLarge]
+  /// [TextStyle], with color set to the app bar's [foregroundColor].
+  /// {@endtemplate}
+  ///
+  /// See also:
+  ///
+  ///  * [toolbarTextStyle], which is the default text style for the AppBar's
+  ///    [title], [leading], and [actions] widgets, also known as the
+  ///    AppBar's "toolbar".
+  ///  * [DefaultTextStyle], which overrides the default text style for all of the
+  ///    widgets in a subtree.
+  final TextStyle? titleTextStyle;
 
-  void _handleDrawerButtonEnd() {
-    Scaffold.of(context).openEndDrawer();
-  }
+  /// {@template flutter.material.appbar.systemOverlayStyle}
+  /// Specifies the style to use for the system overlays (e.g. the status bar on
+  /// Android or iOS, the system navigation bar on Android).
+  ///
+  /// If this property is null, then [AppBarTheme.systemOverlayStyle] of
+  /// [ThemeData.appBarTheme] is used. If that is also null, an appropriate
+  /// [SystemUiOverlayStyle] is calculated based on the [backgroundColor].
+  ///
+  /// The AppBar's descendants are built within a
+  /// `AnnotatedRegion<SystemUiOverlayStyle>` widget, which causes
+  /// [SystemChrome.setSystemUIOverlayStyle] to be called
+  /// automatically. Apps should not enclose an AppBar with their
+  /// own [AnnotatedRegion].
+  /// {@endtemplate}
+  //
+  /// See also:
+  ///
+  ///  * [AnnotatedRegion], for placing [SystemUiOverlayStyle] in the layer tree.
+  ///  * [SystemChrome.setSystemUIOverlayStyle], the imperative API for setting
+  ///    system overlays style.
+  final SystemUiOverlayStyle? systemOverlayStyle;
+
+  /// {@template flutter.material.appbar.forceMaterialTransparency}
+  /// Forces the AppBar's Material widget type to be [MaterialType.transparency]
+  /// (instead of Material's default type).
+  ///
+  /// This will remove the visual display of [backgroundColor] and [elevation],
+  /// and affect other characteristics of the AppBar's Material widget.
+  ///
+  /// Provided for cases where the app bar is to be transparent, and gestures
+  /// must pass through the app bar to widgets beneath the app bar (i.e. with
+  /// [Scaffold.extendBodyBehindAppBar] set to true).
+  ///
+  /// Defaults to false.
+  /// {@endtemplate}
+  final bool forceMaterialTransparency;
+
+  /// {@macro flutter.material.Material.clipBehavior}
+  final Clip? clipBehavior;
+
+  const ContextualAppBar({
+    super.key,
+    required this.counterBuilder,
+    required this.contextualActions,
+    this.automaticallyImplyLeading = true,
+    this.closeIcon = Icons.close,
+    this.leadingWidth,
+    this.flexibleSpace,
+    this.bottom,
+    this.elevation,
+    this.scrolledUnderElevation,
+    this.notificationPredicate = defaultScrollNotificationPredicate,
+    this.shadowColor,
+    this.surfaceTintColor,
+    this.shape,
+    this.backgroundColor,
+    this.foregroundColor,
+    this.iconTheme,
+    this.actionsIconTheme,
+    this.primary = true,
+    this.centerTitle,
+    this.excludeHeaderSemantics = false,
+    this.titleSpacing,
+    this.toolbarOpacity = 1.0,
+    this.bottomOpacity = 1.0,
+    this.toolbarHeight,
+    this.toolbarTextStyle,
+    this.titleTextStyle,
+    this.systemOverlayStyle,
+    this.forceMaterialTransparency = false,
+    this.clipBehavior,
+  });
 
   @override
   Widget build(BuildContext context) {
-    assert(!widget.primary || debugCheckHasMediaQuery(context));
-    assert(debugCheckHasMaterialLocalizations(context));
-    final ThemeData theme = Theme.of(context);
-    final AppBarTheme appBarTheme = AppBarTheme.of(context);
-    final ScaffoldState scaffold = Scaffold.of(context);
-
-    final bool hasEndDrawer = scaffold.hasEndDrawer;
-
-    IconThemeData overallIconTheme =
-        widget.iconTheme ?? appBarTheme.iconTheme ?? theme.primaryIconTheme;
-    IconThemeData actionsIconTheme = widget.actionsIconTheme ??
-        appBarTheme.actionsIconTheme ??
-        overallIconTheme;
-    TextStyle? centerStyle = widget.textTheme?.headline6 ??
-        appBarTheme.textTheme?.headline6 ??
-        theme.primaryTextTheme.headline6;
-    TextStyle? sideStyle = widget.textTheme?.bodyText2 ??
-        appBarTheme.textTheme?.bodyText2 ??
-        theme.primaryTextTheme.bodyText2;
-
-    if (widget.toolbarOpacity != 1.0) {
-      final double opacity =
-          const Interval(0.25, 1.0, curve: Curves.fastOutSlowIn)
-              .transform(widget.toolbarOpacity);
-      if (centerStyle?.color != null)
-        centerStyle = centerStyle!
-            .copyWith(color: centerStyle.color!.withOpacity(opacity));
-      if (sideStyle?.color != null)
-        sideStyle =
-            sideStyle!.copyWith(color: sideStyle.color!.withOpacity(opacity));
-      overallIconTheme = overallIconTheme.copyWith(
-          opacity: opacity * (overallIconTheme.opacity ?? 1.0));
-      actionsIconTheme = actionsIconTheme.copyWith(
-          opacity: opacity * (actionsIconTheme.opacity ?? 1.0));
-    }
-
-    Widget? actions;
-    if (widget.contextualActions.isNotEmpty) {
-      actions = Row(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: widget.contextualActions,
-      );
-    } else if (hasEndDrawer) {
-      actions = IconButton(
-        icon: const Icon(Icons.menu),
-        onPressed: _handleDrawerButtonEnd,
-        tooltip: MaterialLocalizations.of(context).openAppDrawerTooltip,
-      );
-    }
-
-    // Allow the trailing actions to have their own theme if necessary.
-    if (actions != null) {
-      actions = IconTheme.merge(
-        data: actionsIconTheme,
-        child: actions,
-      );
-    }
-
-    final Widget toolbar = NavigationToolbar(
-      leading: IconButton(
-          icon: Icon(widget.closeIcon ?? Icons.close),
-          onPressed: () => ActionMode.disable<T>(context)),
-      middle: Consumer<ItemsController<T>>(
-        builder:
-            (BuildContext context, ItemsController<T> value, Widget? child) {
-          return DefaultTextStyle(
-            style: centerStyle!,
-            softWrap: false,
-            overflow: TextOverflow.ellipsis,
-            child: Semantics(
-              namesRoute: true,
-              child: _AppBarTitleBox(
-                  child: widget.counterBuilder(value.items.length)),
-              header: true,
-            ),
-          );
-        },
-      ),
-      trailing: actions,
-      centerMiddle: widget._getEffectiveCenterTitle(theme)!,
-      middleSpacing: widget.titleSpacing,
-    );
-
-    // If the toolbar is allocated less than kToolbarHeight make it
-    // appear to scroll upwards within its shrinking container.
-    Widget appBar = ClipRect(
-      child: CustomSingleChildLayout(
-        delegate: const _ToolbarContainerLayout(),
-        child: IconTheme.merge(
-          data: overallIconTheme,
-          child: DefaultTextStyle(
-            style: sideStyle!,
-            child: toolbar,
+    return Consumer<ItemsController<T>>(
+      builder: (BuildContext context, ItemsController<T> value, Widget? child) {
+        return AppBar(
+          leading: IconButton(
+            icon: Icon(closeIcon),
+            onPressed: () {
+              ActionMode.disable<T>(context);
+            },
           ),
-        ),
-      ),
-    );
-    if (widget.bottom != null) {
-      appBar = Column(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: <Widget>[
-          Flexible(
-            child: ConstrainedBox(
-              constraints: const BoxConstraints(maxHeight: kToolbarHeight),
-              child: appBar,
-            ),
-          ),
-          if (widget.bottomOpacity == 1.0)
-            widget.bottom ?? Container()
-          else
-            Opacity(
-              opacity: const Interval(0.25, 1.0, curve: Curves.fastOutSlowIn)
-                  .transform(widget.bottomOpacity),
-              child: widget.bottom,
-            ),
-        ],
-      );
-    }
-
-    // The padding applies to the toolbar and tab bar, not the flexible space.
-    if (widget.primary) {
-      appBar = SafeArea(
-        top: true,
-        child: appBar,
-      );
-    }
-
-    appBar = Align(
-      alignment: Alignment.topCenter,
-      child: appBar,
-    );
-
-    if (widget.flexibleSpace != null) {
-      appBar = Stack(
-        fit: StackFit.passthrough,
-        children: <Widget>[
-          widget.flexibleSpace ?? Container(),
-          appBar,
-        ],
-      );
-    }
-    final Brightness brightness = widget.brightness ??
-        appBarTheme.brightness ??
-        theme.primaryColorBrightness;
-    final SystemUiOverlayStyle overlayStyle = brightness == Brightness.dark
-        ? SystemUiOverlayStyle.light
-        : SystemUiOverlayStyle.dark;
-
-    return Semantics(
-      container: true,
-      child: AnnotatedRegion<SystemUiOverlayStyle>(
-        value: overlayStyle,
-        child: Material(
-          color:
-              widget.backgroundColor ?? appBarTheme.color ?? theme.primaryColor,
-          elevation:
-              widget.elevation ?? appBarTheme.elevation ?? _defaultElevation,
-          shape: widget.shape,
-          child: Semantics(
-            explicitChildNodes: true,
-            child: appBar,
-          ),
-        ),
-      ),
+          elevation: elevation,
+          automaticallyImplyLeading: automaticallyImplyLeading,
+          surfaceTintColor: surfaceTintColor,
+          foregroundColor: foregroundColor,
+          centerTitle: centerTitle,
+          bottom: bottom,
+          backgroundColor: backgroundColor,
+          clipBehavior: clipBehavior,
+          actionsIconTheme: actionsIconTheme,
+          bottomOpacity: bottomOpacity,
+          excludeHeaderSemantics: excludeHeaderSemantics,
+          flexibleSpace: flexibleSpace,
+          actions: contextualActions,
+          title: counterBuilder(value.items.length),
+          forceMaterialTransparency: forceMaterialTransparency,
+          iconTheme: iconTheme,
+          leadingWidth: leadingWidth,
+          notificationPredicate: notificationPredicate,
+          primary: primary,
+          scrolledUnderElevation: scrolledUnderElevation,
+          shadowColor: shadowColor,
+          shape: shape,
+          systemOverlayStyle: systemOverlayStyle,
+          titleSpacing: titleSpacing,
+          titleTextStyle: titleTextStyle,
+          toolbarHeight: toolbarHeight,
+          toolbarOpacity: toolbarOpacity,
+          toolbarTextStyle: toolbarTextStyle,
+        );
+      },
     );
   }
-}
-
-class _AppBarTitleBox extends SingleChildRenderObjectWidget {
-  const _AppBarTitleBox({Key? key, required Widget child})
-      : super(key: key, child: child);
-
-  @override
-  _RenderAppBarTitleBox createRenderObject(BuildContext context) {
-    return _RenderAppBarTitleBox(
-      textDirection: Directionality.of(context),
-    );
-  }
-
-  @override
-  void updateRenderObject(
-      BuildContext context, _RenderAppBarTitleBox renderObject) {
-    renderObject.textDirection = Directionality.of(context);
-  }
-}
-
-class _RenderAppBarTitleBox extends RenderAligningShiftedBox {
-  _RenderAppBarTitleBox({
-    RenderBox? child,
-    TextDirection? textDirection,
-  }) : super(
-            child: child,
-            alignment: Alignment.center,
-            textDirection: textDirection);
-
-  @override
-  void performLayout() {
-    final BoxConstraints innerConstraints =
-        constraints.copyWith(maxHeight: double.infinity);
-    child!.layout(innerConstraints, parentUsesSize: true);
-    size = constraints.constrain(child!.size);
-    alignChild();
-  }
-}
-
-class _ToolbarContainerLayout extends SingleChildLayoutDelegate {
-  const _ToolbarContainerLayout();
-
-  @override
-  BoxConstraints getConstraintsForChild(BoxConstraints constraints) {
-    return constraints.tighten(height: kToolbarHeight);
-  }
-
-  @override
-  Size getSize(BoxConstraints constraints) {
-    return Size(constraints.maxWidth, kToolbarHeight);
-  }
-
-  @override
-  Offset getPositionForChild(Size size, Size childSize) {
-    return Offset(0.0, size.height - childSize.height);
-  }
-
-  @override
-  bool shouldRelayout(_ToolbarContainerLayout oldDelegate) => false;
 }
